Extract DB record translation from databaseStream
