+++ src/lib/databaseStream.js
@@ -4,7 +4,7 @@ import _stream from 'stream'
 import through2 from 'through2'
 import { dbClient, createIndex } from './databaseClient.js'
 
-import { getItemCreated } from './database.js'
+import { convertIngestObjectToDbObject } from './ingest.js'
 
 import logger from './logger.js'
 
@@ -123,43 +123,12 @@ export default async function stream() {
       /** @type {any} */ encoding,
       /** @type {(arg0: null | undefined, arg1: { index: string; id: any; action: string; _retry_on_conflict: number; body: any; } | undefined) => void} */ next
     ) => {
-      let index = ''
-      logger.debug('Data', data)
-      if (data && data.hasOwnProperty('extent')) {
-        index = COLLECTIONS_INDEX
-      } else if (data && data.hasOwnProperty('geometry')) {
-        index = data.collection
-      } else {
+      const record = await convertIngestObjectToDbObject(data)
+      if (!record) {
         // @ts-ignore
         next()
         return
       }
-
-      // remove any hierarchy links in a non-mutating way
-      const hlinks = ['self', 'root', 'parent', 'child', 'collection', 'item', 'items']
-      const links = data.links.filter((/** @type {{ rel: string; }} */ link) => !hlinks.includes(link.rel))
-      const dbDataObject = { ...data, links }
-
-      if (data.hasOwnProperty('properties')) {
-        const now = (new Date()).toISOString()
-
-        const created = (await getItemCreated(data.collection, data.id)) || now
-
-        // @ts-ignore
-        dbDataObject.properties.created = created
-        // @ts-ignore
-        dbDataObject.properties.updated = now
-      }
-
-      // create ES record
-      const record = {
-        index,
-        id: dbDataObject.id,
-        action: 'index',
-        _retry_on_conflict: 3,
-        body: dbDataObject
-      }
-
       next(null, record)
     })
